requirements.txt
# Flask 查詢系統核心依賴套件
# 適用於 Python 3.13.7 Windows Server 2022 離線環境

# Web 框架核心
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1

# Oracle 資料庫驅動 (Thin Mode)
python-oracledb==2.1.1

# Excel 匯出功能
pandas==2.2.3
openpyxl==3.1.5

# HTTP 請求（附件下載 API）
requests==2.32.3

# 生產環境部署
waitress==3.0.2

# 設定管理
python-dotenv==1.2.1
